chore - domain, data 모듈 패키지명 변경
