+++ data/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="co.kr.yapp.growth.data"/>
+<manifest package="com.yapp.growth.data"/>
