--- data/src/main/java/co/kr/yapp/growth/data/api/Api.kt
@@ -1,4 +1,0 @@
-package co.kr.yapp.growth.data.api
-
-interface Api {
-}
