--- data/src/main/java/co/kr/yapp/growth/data/entity/Entity.kt
@@ -1,4 +1,0 @@
-package co.kr.yapp.growth.data.entity
-
-class Entity {
-}
