--- data/src/main/java/co/kr/yapp/growth/data/repository/RepositoryImpl.kt
@@ -1,6 +1,0 @@
-package co.kr.yapp.growth.data.repository
-
-import co.kr.yapp.growth.domain.repository.repository
-
-class RepositoryImpl : repository{
-}
