+++ data/src/main/java/com/yapp/growth/data/api/Api.kt
@@ -1,0 +1,4 @@
+package com.yapp.growth.data.api
+
+interface Api {
+}
