+++ data/src/main/java/com/yapp/growth/data/entity/Entity.kt
@@ -1,0 +1,4 @@
+package com.yapp.growth.data.entity
+
+class Entity {
+}
