+++ data/src/main/java/com/yapp/growth/data/repository/RepositoryImpl.kt
@@ -1,0 +1,6 @@
+package com.yapp.growth.data.repository
+
+import com.yapp.growth.domain.repository.repository
+
+class RepositoryImpl : repository{
+}
