--- domain/src/main/java/co/kr/yapp/growth/domain/entity/entity.kt
@@ -1,4 +1,0 @@
-package co.kr.yapp.growth.domain.entity
-
-class entity {
-}
