--- domain/src/main/java/co/kr/yapp/growth/domain/repository/repository.kt
@@ -1,4 +1,0 @@
-package co.kr.yapp.growth.domain.repository
-
-interface repository {
-}
