--- domain/src/main/java/co/kr/yapp/growth/domain/usecase/useCase.kt
@@ -1,4 +1,0 @@
-package co.kr.yapp.growth.domain.usecase
-
-class useCase {
-}
