+++ domain/src/main/java/com/yapp/growth/domain/entity/entity.kt
@@ -1,0 +1,4 @@
+package com.yapp.growth.domain.entity
+
+class entity {
+}
