+++ domain/src/main/java/com/yapp/growth/domain/repository/repository.kt
@@ -1,0 +1,4 @@
+package com.yapp.growth.domain.repository
+
+interface repository {
+}
