+++ domain/src/main/java/com/yapp/growth/domain/usecase/useCase.kt
@@ -1,0 +1,4 @@
+package com.yapp.growth.domain.usecase
+
+class useCase {
+}
